requirements.txt
Django==2.1
djangorestframework==3.11.0
Faker==4.1.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
text-unidecode==1.3
whitenoise==5.1.0
